Set PostgreSQL encoding via attributes, default to LATIN1.
The default to LATIN1 is for compatibility. The attribute can easily be
overridden to UTF8 and whoever does that will know how to make sure the
appropriate locales are installed. The previous way of installing locales
in this cookbook was severely broken (in some cases it completely mangles
the terminal due to starting an interactive dpkg-reconfigure session).

Installing UTF8 locales on Debian can be done with this cookbook:

 https://github.com/danhart/locale-gen/

Make sure you use the latest version, not the one from Opscode Community.
For English, you'd need to use these attributes:

 default[:localegen][:lang] = ['en_US.UTF-8 UTF-8', 'en_US ISO-8859-1']